backend/requirements.txt
fastapi
uvicorn
pandas
mysql-connector-python
ollama
python-dotenv
aiofiles
python-multipart